import a large chunk of msm8916 + msm_shared and associated modules
mostly the changes were mechanical: updating includes, fixing up
makefiles, etc.

the changes to mmu mappings in platform.c is pretty major though, and
is probably completely wrong

for now I opted to hack core headers with the stuff that downstream
expects, if I can actually get this to link and run then I'll figure out
how to clean this up in such a way that it has any hope of going
upstream.